# frontend/openapi.yaml
openapi: 3.1.0
info:
  title: Timverse
  version: 0.1.0
paths:
...
  /api/v1/users/me:
    get:
      tags:
        - users
      summary: Read User Me
      description: Get current user.
      operationId: users-read_user_me
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
      security:
        - OAuth2PasswordBearer: []
    delete:
      tags:
        - users
      summary: Delete User Me
      description: Delete own user.
      operationId: users-delete_user_me
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
      security:
        - OAuth2PasswordBearer: []
    patch:
      tags:
        - users
      summary: Update User Me
      description: Update own user.
      operationId: users-update_user_me
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateMe"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        - OAuth2PasswordBearer: []
  /api/v1/users/me/password:
    patch:
      tags:
        - users
      summary: Update Password Me
      description: Update own password.
      operationId: users-update_password_me
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePassword"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        - OAuth2PasswordBearer: []
  /api/v1/users/:
    get:
      tags:
        - users
      summary: Read Users
      description: Retrieve users.
      operationId: users-read_users
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            default: 0
            title: Skip
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 100
            title: Limit
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersPublic"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    post:
      tags:
        - users
      summary: Create User
      description: Create new user.
      operationId: users-create_user
      security:
        - OAuth2PasswordBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/v1/users/{user_id}:
    get:
      tags:
        - users
      summary: Read User By Id
      description: Get a specific user by id.
      operationId: users-read_user_by_id
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: User Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    patch:
      tags:
        - users
      summary: Update User
      description: Update a user.
      operationId: users-update_user
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: User Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
    delete:
      tags:
        - users
      summary: Delete User
      description: Delete a user.
      operationId: users-delete_user
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            title: User Id
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/v1/users/signup:
    post:
      tags:
        - users
      summary: Register User
      description: Create new user without the need to be logged in.
      operationId: users-register_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegister"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
  /api/v1/users/verify-email/:
    post:
      tags:
        - users
      summary: Verify Email
      description: verify email and reset password.
      operationId: users-verify_email
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewAccount"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "400":
          description: Invalid token or inactive user
        "404":
          description: The user with this email does not exist in the system.
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
...
  /api/v1/private/users/:
    get:
      tags:
        - private
      summary: Get All Users
      description: Get all users.
      operationId: private-get_all_users
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/UserPublic"
                type: array
                title: Response Private-Get All Users
      security:
        - OAuth2PasswordBearer: []
    post:
      tags:
        - private
      summary: Create User
      description: Create a new user.
      operationId: private-create_user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PrivateUserCreate"
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPublic"
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        - OAuth2PasswordBearer: []
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Message:
      properties:
        message:
          type: string
          title: Message
      type: object
      required:
        - message
      title: Message
    NewAccount:
      properties:
        token:
          type: string
          title: Token
      type: object
      required:
        - token
      title: NewAccount
...
    PrivateUserCreate:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
        full_name:
          type: string
          title: Full Name
        is_verified:
          type: boolean
          title: Is Verified
          default: false
      type: object
      required:
        - email
        - password
        - full_name
      title: PrivateUserCreate
...
    UpdatePassword:
      properties:
        current_password:
          type: string
          maxLength: 64
          minLength: 8
          title: Current Password
        new_password:
          type: string
          maxLength: 64
          minLength: 8
          title: New Password
      type: object
      required:
        - current_password
        - new_password
      title: UpdatePassword
    UserCreate:
      properties:
        email:
          type: string
          format: email
          title: Email
        is_active:
          type: boolean
          title: Is Active
        is_superuser:
          type: boolean
          title: Is Superuser
        full_name:
          anyOf:
            - type: string
            - type: "null"
          title: Full Name
        password:
          type: string
          maxLength: 64
          minLength: 8
          title: Password
      type: object
      required:
        - email
        - is_active
        - is_superuser
        - password
      title: UserCreate
    UserPublic:
      properties:
        email:
          type: string
          format: email
          title: Email
        is_active:
          type: boolean
          title: Is Active
        is_superuser:
          type: boolean
          title: Is Superuser
        full_name:
          anyOf:
            - type: string
            - type: "null"
          title: Full Name
        id:
          type: string
          format: uuid
          title: Id
      type: object
      required:
        - email
        - is_active
        - is_superuser
        - id
      title: UserPublic
    UserRegister:
      properties:
        email:
          type: string
          maxLength: 255
          format: email
          title: Email
        password:
          type: string
          maxLength: 64
          minLength: 8
          title: Password
        full_name:
          anyOf:
            - type: string
              maxLength: 255
            - type: "null"
          title: Full Name
      type: object
      required:
        - email
        - password
      title: UserRegister
    UserUpdate:
      properties:
        is_active:
          anyOf:
            - type: boolean
            - type: "null"
          title: Is Active
        is_superuser:
          anyOf:
            - type: boolean
            - type: "null"
          title: Is Superuser
        full_name:
          anyOf:
            - type: string
            - type: "null"
          title: Full Name
        email:
          anyOf:
            - type: string
              maxLength: 255
              format: email
            - type: "null"
          title: Email
      type: object
      title: UserUpdate
    UserUpdateMe:
      properties:
        is_active:
          anyOf:
            - type: boolean
            - type: "null"
          title: Is Active
        full_name:
          anyOf:
            - type: string
              maxLength: 255
            - type: "null"
          title: Full Name
        email:
          anyOf:
            - type: string
              maxLength: 255
              format: email
            - type: "null"
          title: Email
      type: object
      title: UserUpdateMe
    UsersPublic:
      properties:
        data:
          items:
            $ref: "#/components/schemas/UserPublic"
          type: array
          title: Data
        count:
          type: integer
          title: Count
      type: object
      required:
        - data
        - count
      title: UsersPublic
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/login/access-token